requirements.txt
# Required dependencies
PyQt6>=6.0.0
numpy>=1.20.0
pyserial>=3.5
pyyaml>=6.0
jsonschema>=4.0.0

# Optional dependencies
matplotlib>=3.5.0 # For visualization
pandas>=1.3.0     # For data processing
